use only optimized discovery since js-navigation out of scope
