+++ tests/org.eclipse.n4js.cli.tests/src/org/eclipse/n4js/cli/projectdiscovery/tests/ProjectDiscoveryTest.java
@@ -10,15 +10,21 @@
  */
 package org.eclipse.n4js.cli.projectdiscovery.tests;
 
+import static org.junit.Assert.assertEquals;
+
 import java.io.File;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.LinkedHashSet;
 
 import org.eclipse.n4js.cli.N4jscFactory;
+import org.eclipse.n4js.cli.projectdiscovery.tests.CreateProjectStructureUtils.ProjectDiscoveryTestData;
 import org.eclipse.n4js.utils.ProjectDiscoveryHelper;
+import org.eclipse.n4js.utils.io.FileUtils;
 import org.junit.BeforeClass;
+import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
@@ -26,19 +32,23 @@ import org.junit.runners.Parameterized.Parameters;
 import com.google.inject.Injector;
 
 /**
- * Test checks if {@link ProjectDiscoveryHelper} find all projects and their dependencies.
+ * Abstract test to find all projects and their dependencies.
  */
 @RunWith(Parameterized.class)
-public class ProjectDiscoveryTest extends AbstractProjectDiscoveryTest {
+public class ProjectDiscoveryTest {
 	static ProjectDiscoveryHelper projectDiscoveryHelper;
 
 	/** Find test data files */
-	@Parameters(name = "{index}: {0}")
+	@Parameters(name = "{0}")
 	public static Collection<File> testData() {
-		Collection<File> tests = new ArrayList<>();
-		tests.addAll(getPDTFilesIn(new File("DiscoveryTestsBoth")));
-		tests.addAll(getPDTFilesIn(new File("DiscoveryTestsOnlyNonOptimized")));
-		return tests;
+		File folder = new File("ProjectDiscoveryTest");
+		ArrayList<File> testFiles = new ArrayList<>();
+		for (String testFileName : folder.list()) {
+			if (testFileName.endsWith(".pdt")) {
+				testFiles.add(new File(folder, testFileName));
+			}
+		}
+		return testFiles;
 	}
 
 	/** Init the {@link ProjectDiscoveryHelper} */
@@ -48,13 +58,36 @@ public class ProjectDiscoveryTest extends AbstractProjectDiscoveryTest {
 		projectDiscoveryHelper = injector.getInstance(ProjectDiscoveryHelper.class);
 	}
 
+	final File testFile;
+
 	/** Constructor */
 	public ProjectDiscoveryTest(File testFile) {
-		super(testFile);
+		this.testFile = testFile;
 	}
 
-	@Override
-	protected LinkedHashSet<Path> callProjectDiscoveryHelper(Path workspaceRoot) {
-		return projectDiscoveryHelper.collectAllProjectDirs(workspaceRoot);
+	/** Execute parameterized test */
+	@Test
+	public void test() {
+		String fileName = testFile.getName();
+		Path tmpDir = FileUtils.createTempDirectory(fileName + "_");
+
+		ProjectDiscoveryTestData pdtd = CreateProjectStructureUtils.readPDTFile(testFile);
+		CreateProjectStructureUtils.createFolderStructure(tmpDir.toFile(), pdtd);
+
+		Path workspaceRoot = new File(tmpDir.toFile(), pdtd.workingDir.getPath()).toPath();
+
+		ArrayList<String> actualFolders = getActualFolders(tmpDir, workspaceRoot);
+		assertEquals(pdtd.expectedProjects, actualFolders);
+	}
+
+	private ArrayList<String> getActualFolders(Path tmpDir, Path workspaceRoot) {
+		LinkedHashSet<Path> projectDirs = projectDiscoveryHelper.collectAllProjectDirs(workspaceRoot);
+		ArrayList<String> actualFolders = new ArrayList<>();
+		for (Path projectDir : projectDirs) {
+			String relativeFolder = tmpDir.relativize(projectDir).toString();
+			actualFolders.add(relativeFolder);
+		}
+		Collections.sort(actualFolders);
+		return actualFolders;
 	}
 }
